Add rule for making a snakemake profile
The main purpose of generating a profile is to handle the configuration
of a cluster environment. This is related to issue #4, but I don't think
I'm ready to close it just yet. I will have to add some safeguards when
cloning the cookicutter repository, and also change some wordings in the
README and in the config. The command for this rule should probably also
be made into a standalone script to make the snakefile less messy.